refactor: keep CommonJS compat for module
GoogleAuth is being exported as an ES6 class. This breaks compatibility
with existing JS users of this library for the main module export. This
change adds an index.js that re-exports GoogleAuth as a function.

Note that we are not using this hack for every single file in this
module. Anyone reaching into the internals of this module will need to
update their code.