chore: docs and samples for refresh token, update uris (#215)
Note: This is a semver minor change (though it will be included in the 1.0 release). I have no reason to believe it breaks any functionality, but it does (I think) bring us up to date with the most recent documentation:
https://developers.google.com/identity/protocols/OAuth2InstalledApp

It's also suggested in #120 that this is causing problems for users as well.