feat(refresh): add forceRefreshOnFailure flag
It's highly unlikely, but the expiry_date could be wrong, or the
comparison in isTokenExpiring could fail due to an inaccurate system
clock. As a further fallback behavior, some developers may wish to opt-
in to attempting a refresh if Google's servers respond with a status
code that may indicate the token has expired.